requirements.txt
boto3
botocore
pdfplumber
pandas
numpy
python-dotenv
requests
